% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsim.R
\name{microsim}
\alias{microsim}
\title{Demographic Microsimulation}
\usage{
microsim(
  initial_popn,
  fertility_rates,
  mortality_rates,
  step = 5L,
  dominant_sex = "Female"
)
}
\arguments{
\item{initial_popn}{An array with dimensions "age",
and "sex".}

\item{fertility_rates}{An array with dimensions
"age", "triangle", "sex", and "time".}

\item{mortality_rates}{An array with dimensions
"age", "triangle", "sex", and "time".}

\item{step}{The length of one age-time step.
Defaults to 5.}

\item{dominant_sex}{Which sex provides the exposure
term for birth rates. Defaults to "Female".}
}
\value{
A list of arrays with names
"population", "births", and "deaths".
}
\description{
Given an initial population, and a set of
fertility and mortality rates,
randomly generate subsequent population, births,
and deaths.
}
\details{
All dimensions of \code{initial_popn}, \code{fertility_rates},
and \code{mortality_rates} with the same
name should have the same length:
for instance, all dimensions called "age" should have the same
length. This includes the "age" dimension of \code{fertility_rates}.
Rates outside the reproductive ages should be set to zero.

The "triangle" dimension holds Lexis triangles, and
must have labels "Lower" and "Upper".

The "sex" dimension can be length 1, 2, or more.
}
