% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallpopn-package.R
\docType{package}
\name{smallpopn-package}
\alias{smallpopn}
\alias{smallpopn-package}
\title{Package 'smallpopn'}
\description{
Functions needed for paper "How big does a population
need to be before demographers can ignore individual-level
randomness in demographic events?"
}
\section{Functions}{

\itemize{
\item \code{Lx} - Life table values for West model life table.
\item \code{microsim} - Function to carry out demographic microsimulation.
\item \code{mx} - Mortality rates for West model life table.
\item \code{propn_age_fert} - Age distribution of fertility from Booth standard
}
}

\author{
\strong{Maintainer}: John Bryant \email{john@bayesiandemography.com} (\href{https://orcid.org/0000-0002-8010-9748}{ORCID})

}
\keyword{internal}
